Changes to be committed: 	modified:   src/Python PYSAT_GUI/PYSAT_UI_MODULES/file_outpath_.py 	modified:   src/Python PYSAT_GUI/PYSAT_UI_MODULES/get_data_.py 	modified:   src/Python PYSAT_GUI/PYSAT_UI_MODULES/get_mask_.py 	modified:   src/Python PYSAT_GUI/PYSAT_UI_MODULES/pca_.py 	modified:   src/Python PYSAT_GUI/PYSAT_UI_MODULES/pls_.py 	modified:   src/Python PYSAT_GUI/PYSAT_UI_MODULES/ransac_.py 	modified:   src/Python PYSAT_GUI/PYSAT_UI_MODULES/regression_.py 	modified:   src/Python PYSAT_GUI/UnitTests.py 	renamed:    src/Python PYSAT_GUI/pysat_func.py -> src/Python PYSAT_GUI/pysat_function.py 	modified:   src/Python PYSAT_GUI/pysat_ui.py
