+++ src/Python PYSAT_GUI/pysat_ui.py
@@ -1,8 +1,6 @@
 from PyQt4 import QtCore, QtGui
-import pysat_func
+from pysat_function import pysat_func
 import PYSAT_UI_MODULES
-
-import pysat
 
 try:
     _fromUtf8 = QtCore.QString.fromUtf8
@@ -19,11 +17,6 @@ try:
 except AttributeError:
     def _translate(context, text, disambig):
         return QtGui.QApplication.translate(context, text, disambig)
-
-
-
-
-
 
 class pysat_ui(object):
     def __init__(self):
